+++ src/main/java/hr/dlatecki/algorithms/gen_alg/codecs/abstracts/package-info.java
@@ -1,6 +1,9 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains abstract implementations of several codec types.
+ * Contains abstract implementations of several codec types.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
  */
 package hr.dlatecki.algorithms.gen_alg.codecs.abstracts;