+++ src/main/java/hr/dlatecki/algorithms/gen_alg/codecs/package-info.java
@@ -1,0 +1,10 @@
+// Copyright (C) 2016 Domagoj Latečki
+/* You may use, distribute and modify this code under the terms of the MIT license. */
+/**
+ * Contains sub-packages which define interfaces and abstract classes for codecs. Also contains several implementations
+ * of the codecs.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg.codecs;