+++ src/main/java/hr/dlatecki/algorithms/gen_alg/exceptions/package-info.java
@@ -1,7 +1,7 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains exceptions which are used in genetic algorithms, operators and population.
+ * Contains exceptions which are used in genetic algorithms, operators and population.
  * 
  * @author Domagoj Latečki
  * @since 1.8