+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/interfaces/package-info.java
@@ -1,7 +1,7 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains interfaces that define operators in genetic algorithms.
+ * Contains interfaces that define operators in genetic algorithms.
  * 
  * @author Domagoj Latečki
  * @since 1.8