+++ src/main/java/hr/dlatecki/algorithms/gen_alg/package-info.java
@@ -1,8 +1,8 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains implementations of the genetic algorithms. It also contains sub-packages which define
- * population and operators in the algorithms.
+ * Contains implementations of the genetic algorithms. It also contains sub-packages which define population and
+ * operators in the algorithms.
  * 
  * @author Domagoj Latečki
  * @since 1.8