+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/abstracts/package-info.java
@@ -1,7 +1,7 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains abstract implementations of the interfaces defined in
+ * Contains abstract implementations of the interfaces defined in
  * {@link hr.dlatecki.algorithms.gen_alg.population.interfaces} package.
  * 
  * @author Domagoj Latečki