+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/chromosomes/package-info.java
@@ -1,7 +1,7 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Package which contains implementations of the chromosomes used in genetic algorithms defined in
+ * Contains implementations of the chromosomes used in genetic algorithms defined in
  * {@link hr.dlatecki.algorithms.gen_alg} package.
  * 
  * @author Domagoj Latečki