+++ src/test/java/hr/dlatecki/algorithms/gen_alg/test_utils/package-info.java
@@ -1,9 +1,9 @@
 // Copyright (C) 2016 Domagoj Latečki
 /* You may use, distribute and modify this code under the terms of the MIT license. */
 /**
- * Contains interfaces that define population and fitness tests.
+ * Contains classes which provide methods and constants in testing, or are used for mocking other classes.
  * 
  * @author Domagoj Latečki
  * @since 1.8
  */
-package hr.dlatecki.algorithms.gen_alg.population.interfaces;
+package hr.dlatecki.algorithms.gen_alg.test_utils;